% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_bumphunter.R
\name{DMRInt_bumphunter}
\alias{DMRInt_bumphunter}
\title{Identify the differentially methylated regions using bumphunter method.
This function identifies the differentially methylated regions by using bumphunter method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of probes in a DMR.}
\usage{
DMRInt_bumphunter(beta, pheno, arraytype = c("450K", "EPIC"), minProbes)
}
\arguments{
\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{arraytype}{The type of methylation array, including methylation 450K and EPIC array.}

\item{minProbes}{The threshold for minimum number of probes in a DMR.}

\item{pd}{The phenomenon data with sample information (With Sample group column named Sample_Group).}
}
\value{
The DMR set detected by bumphunter.
}
\description{
Identify the differentially methylated regions using bumphunter method.
This function identifies the differentially methylated regions by using bumphunter method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of probes in a DMR.
}
\examples{
beta = readRDS(system.file("extdata","beta.RDS",package = 'DMRIntTk'))
pd=read.csv(system.file("extdata","pd.csv",package = 'DMRIntTk'))
bumphunter=DMRInt_bumphunter(beta = beta, pheno=pd$Sample_Group, arraytype = "450K", minProbes = 3)

}
