% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_densitypeak.R
\name{DMRInt_densitypeak}
\alias{DMRInt_densitypeak}
\title{Cluster all intervals based on density peak algorithm.}
\usage{
DMRInt_densitypeak(
  interval,
  totalDMR,
  prefer = c("probe", "length"),
  arraytype = c("450K", "EPIC")
)
}
\arguments{
\item{interval}{The interval with calculated weights(obtained from DMRInt_weight function).}

\item{totalDMR}{The DMR sets obtained from DMRIntTk_input function.}
}
\value{
The integrated DMR set.
}
\description{
Cluster all intervals based on density peak algorithm.
}
\examples{
Res=DMRInt_densitypeak(interval_weight, totalDMR, prefer = "probe", arraytype = "450K")
}
