% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_ipDMR.R
\name{DMRInt_ipDMR}
\alias{DMRInt_ipDMR}
\title{Identify the differentially methylated regions using ipDMR method.
This function identifies the differentially methylated regions by using ipDMR method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of CpGs in a DMR.}
\usage{
DMRInt_ipDMR(beta, pd, arraytype = c("450K", "EPIC"), minProbes)
}
\arguments{
\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{pd}{The phenomenon data file with sample information (With Sample group column named Sample_Group).}

\item{arraytype}{The type of methylation array, including 450K and EPIC array.}

\item{minProbes}{The threshold for minimum number of probes in a DMR.}
}
\value{
The DMR set detected by ipDMR.
}
\description{
Identify the differentially methylated regions using ipDMR method.
This function identifies the differentially methylated regions by using ipDMR method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of CpGs in a DMR.
}
\examples{
beta = readRDS(system.file("extdata","beta.RDS",package = 'DMRIntTk'))
pd=read.csv(system.file("extdata","pd.csv",package = 'DMRIntTk'))
ipDMR=DMRInt_ipDMR(beta = beta, pd = pd, arraytype = "450K", minProbes = 3)


}
