% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_matrix.R
\name{DMRInt_matrix}
\alias{DMRInt_matrix}
\title{Construct the reliability weight matrix.
This function calculates the DMRscore and weights of all methods under different methylation difference thresholds and constructs the reliability weight matrix for each method.}
\usage{
DMRInt_matrix(totalDMR)
}
\arguments{
\item{totalDMR}{DMRs sets obtained from the function preinput.}
}
\value{
The reliability weight matrix.
}
\description{
Construct the reliability weight matrix.
This function calculates the DMRscore and weights of all methods under different methylation difference thresholds and constructs the reliability weight matrix for each method.
}
\examples{
weight_m = DMRInt_matrix(totalDMR)
}
