% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_mCSEA.R
\name{DMRInt_mCSEA}
\alias{DMRInt_mCSEA}
\title{Identify the differentially methylated regions using mCSEA method.
This function identifies the differentially methylated regions by using mCSEA method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the caseGroup
and the refGroup, the interested regions types, the type of platform, and the threshold for minimum number of CpGs in a DMR.}
\usage{
DMRInt_mCSEA(
  beta,
  pd,
  caseGroup,
  refGroup,
  regionsTypes = c("promoters", "genes", "CGI"),
  platform = c("450k", "EPIC"),
  minCpGs
)
}
\arguments{
\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{pd}{The phenomenon data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAWElEQVR42mNgGPTAxsZmJsVqQApgmGw1yApwKcQiT7phRBuCzzCSDSHGMKINIeDNmWQlA2IigKJwIssQkHdINgxfmBBtGDEBS3KCxBc7pMQgMYE5c/AXPwAwSX4lV3pTWwAAAABJRU5ErkJggg==file with sample information (With Sample group column named Sample_Group).}

\item{caseGroup}{The name of case group in the pd file.}

\item{refGroup}{The name of control group in the pd file.}

\item{regionsTypes}{The region types you are interested in. mCSEA will identify the DMRs in that specific type of regions.}

\item{platform}{The type of methylation array, including 450K and EPIC array.}

\item{minCpGs}{The threshold for minimum number of CpGs in a DMR.}
}
\value{
The DMR set detected by mCSEA.
}
\description{
Identify the differentially methylated regions using mCSEA method.
This function identifies the differentially methylated regions by using mCSEA method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the caseGroup
and the refGroup, the interested regions types, the type of platform, and the threshold for minimum number of CpGs in a DMR.
}
\examples{
beta = readRDS(system.file("extdata","beta.RDS",package = 'DMRIntTk'))
pd=read.csv(system.file("extdata","pd.csv",package = 'DMRIntTk'))
mCSEA=DMRInt_mCSEA(beta, pd, caseGroup= "Tumor", refGroup = "Normal", regionsTypes = "promoters", platform = "450k", minCpGs = 3)

}
