% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_method.R
\name{DMRInt_method}
\alias{DMRInt_method}
\title{Determine DMRs that cover each bin, and calculate the numbers of methods that cover the bin.}
\usage{
DMRInt_method(input_DMR, arraytype = c("450K", "EPIC"))
}
\arguments{
\item{input_DMR}{DMRs sets obtained from the function DMRInt_input.}

\item{arraytype}{The type of methylation array, including 450K and EPIC array.}
}
\value{
The bin with covered DMRs and methods.
}
\description{
Determine DMRs that cover each bin, and calculate the numbers of methods that cover the bin.
}
\examples{
bin_method = DMRInt_method(input_DMR, arraytype = "450K" )
}
