% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_seqlm.R
\name{DMRInt_seqlm}
\alias{DMRInt_seqlm}
\title{Identify the differentially methylated regions using seqlm method.
This function identifies the differentially methylated regions by using seqlm method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of probes in a DMR.}
\usage{
DMRInt_seqlm(beta, pd, arraytype = "450K", minCpGs)
}
\arguments{
\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{pd}{The phenomenon data with sample information (With Sample group column named Sample_Group).}

\item{arraytype}{The type of methylation array, 450K is allowed.}

\item{minProbes}{The threshold for minimum number of probes in a DMR.}
}
\value{
The DMR set detected by seqlm.
}
\description{
Identify the differentially methylated regions using seqlm method.
This function identifies the differentially methylated regions by using seqlm method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of probes in a DMR.
}
\examples{
beta = readRDS(system.file("extdata","beta.RDS",package = 'DMRIntTk'))
pd=read.csv(system.file("extdata","pd.csv",package = 'DMRIntTk'))
seqlm=DMRInt_seqlm(beta = beta, pd = pd, arraytype = "450K", minCpGs = 3)

}
