% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_weight.R
\name{DMRInt_weight}
\alias{DMRInt_weight}
\title{Calculate the weight of intervals.}
\usage{
DMRInt_weight(interval, weight_m, beta, case, control)
}
\arguments{
\item{interval}{The pre-split genomic intervals with calculated covered DMRs(obtained from DMRIntTk_method function).}

\item{weight_m}{The weight matrix obtained from DMRInt_matrix function.}
}
\value{
The interval with calculated weights.
}
\description{
Calculate the weight of intervals.
}
\examples{
interval_weight=DMRInt_weight(interval_method,weight_m,beta,case,control)
}
