% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_DMR.R
\name{identify_DMR}
\alias{identify_DMR}
\title{Identify the differentially methylated regions using bumphunter method.
This function identifies the differentially methylated regions by using bumphunter method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of probes in a DMR.}
\usage{
identify_DMR(
  beta = beta,
  pheno = pd,
  method = c("bumphunter", "combp", "ipDMR", "mCSEA", "ProbeLasso", "seqlm"),
  arraytype = c("450K", "EPIC"),
  group1 = group1,
  group2 = group2,
  minProbes = minProbes,
  regionsTypes = c("promoters", "genes", "CGI")
)
}
\arguments{
\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{method}{The input DMR identification methods, including bumphunter, comb-p, ipDMR, mCSEA, ProbeLasso, and seqlm. One or more methods are allowed to be used.}

\item{arraytype}{The type of methylation array, including methylation 450K and EPIC array.}

\item{group1}{The group name of group one.}

\item{group2}{The group name of group two.}

\item{minProbes}{The threshold for minimum number of probes in a DMR.}

\item{regionsTypes}{The predefined regions to be analyzed. NULL to skip this step and use customAnnotation (This parameter is only required by mCSEA).}

\item{pd}{The phenomenon data of sample information (With Sample group column named Sample_Group).}
}
\value{
The DMR set detected by input DMR identification methods.
}
\description{
Identify the differentially methylated regions using bumphunter method.
This function identifies the differentially methylated regions by using bumphunter method. Users should provide corresponding
file and parameters, including the methylation beta matrix file, the phenomenon file with sample information, the type of platform, and the threshold for minimum number of probes in a DMR.
}
\examples{
## For 450K data:
beta = load(system.file("extdata", "beta_450K.RData", package = 'DMRIntTk'))
pd = read.csv(system.file("extdata", "pd_450K.csv", package = 'DMRIntTk'))
totalDMR = identify_DMR(beta = beta, method = c("bumphunter","combp","ipDMR","mCSEA","ProbeLasso","seqlm"), pheno = pd, arraytype = "450K", group1 = "Tumor", group2 = "Normal", minProbes = 3, regionsTypes = "promoter")

##For EPIC data:
beta = load(system.file("extdata", "beta_EPIC.RData", package = 'DMRIntTk'))
pd = read.csv(system.file("extdata", "pd_EPIC.csv", package = 'DMRIntTk'))
totalDMR = identify_DMR(beta = beta, method = c("bumphunter","combp","ipDMR","mCSEA","seqlm"), pheno = pd, arraytype = "EPIC", group1 = "CRC", group2 = "Polyp", minProbes = 3)
}
