% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_input.R
\name{DMRInt_input}
\alias{DMRInt_input}
\title{Prepare the input of DMRIntTk.
This function finds the probes included in each DMR, and calculates the methylation differences of probes and DMRs.}
\usage{
DMRInt_input(
  totalDMR,
  beta,
  case_pos = case,
  con_pos = control,
  arraytype = c("450K", "EPIC")
)
}
\arguments{
\item{totalDMR}{Multiple DMR sets predicted from different methods.}

\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{arraytype}{The type of methylation array, including 450K and EPIC array.}

\item{case}{The column of case group.}

\item{control}{The column of control group.}
}
\value{
DMR sets matrix with calculated methylation differences and probes.
}
\description{
Prepare the input of DMRIntTk.
This function finds the probes included in each DMR, and calculates the methylation differences of probes and DMRs.
}
\examples{
If you have used the DMR detection functions provided by DMRIntTk and
obtained different DMR sets, you can combine them to get the total DMR set:
DMRstring = c("chr","start","end", "methodname")
totalDMR = rbind(bumphunter[,DMRstring], ProbeLasso[,DMRstring], mCSEA[,DMRstring], seqlm[,DMRstring], combp[,DMRstring], ipDMR[,DMRstring])

If you already have the total DMR sets matrix, or want to use our example DMR matrix directly:
totalDMR = read.csv(system.file("extdata","totalDMR.csv",package = 'DMRIntTk'))
beta = readRDS(system.file("extdata","beta.RDS",package = 'DMRIntTk'))
case = grep("Tumor",colnames(beta))
control = grep("Normal",colnames(beta))
totalDMR=DMRInt_input(totalDMR, beta, case, control, arraytype = "450K")


}
