% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_input.R
\name{DMRInt_input}
\alias{DMRInt_input}
\title{Prepare the input of DMRIntTk.
This function finds the probes included in each DMR, and calculates the methylation differences of probes and DMRs.}
\usage{
DMRInt_input(
  totalDMR,
  beta,
  group1 = group1,
  group2 = group2,
  arraytype = c("450K", "EPIC")
)
}
\arguments{
\item{totalDMR}{Multiple DMR sets predicted from different methods.}

\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{group1}{The group name of group one.}

\item{group2}{The group name of group two.}

\item{arraytype}{The type of methylation array, including 450K and EPIC array.}
}
\value{
DMR sets matrix with calculated methylation differences and probes.
}
\description{
Prepare the input of DMRIntTk.
This function finds the probes included in each DMR, and calculates the methylation differences of probes and DMRs.
}
\examples{
For 450K data:
input_DMR = DMRInt_input(totalDMR, beta , group1 = "Tumor", group2 = "Normal" , arraytype = "450K")

For EPIC data:
input_DMR = DMRInt_input(totalDMR, beta , group1 = "CRC", group2 = "Polyp" , arraytype = "EPIC")
}
