% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_method.R
\name{DMRInt_method}
\alias{DMRInt_method}
\title{Determine DMRs that cover each interval, and calculate the numbers of methods that cover the interval.}
\usage{
DMRInt_method(totalDMR, arraytype = c("450K", "EPIC"))
}
\arguments{
\item{totalDMR}{DMRs sets obtained from the function DMRInt_input.}

\item{arraytype}{The type of methylation array, including 450K and EPIC array.}
}
\value{
The interval with covered DMRs and methods.
}
\description{
Determine DMRs that cover each interval, and calculate the numbers of methods that cover the interval.
}
\examples{
interval_method = DMRInt_method(totalDMR, arraytype = "450K" )
}
