% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRInt_weight.R
\name{DMRInt_weight}
\alias{DMRInt_weight}
\title{Calculate the weight of intervals.}
\usage{
DMRInt_weight(bin, input_DMR, pd, beta, group1, group2)
}
\arguments{
\item{bin}{The pre-split genomic intervals with calculated covered DMRs (obtained from DMRInt_method function).}

\item{input_DMR}{The DMR set (obtained from the function DMRInt_input function).}

\item{pd}{The phenomenon data of sample information (With Sample group column named Sample_Group).}

\item{beta}{The methylation beta values matrix of samples from two groups.}

\item{group1}{The group name of group one.}

\item{group2}{The group name of group two.}
}
\value{
The bins with calculated weights.
}
\description{
Calculate the weight of intervals.
}
\examples{
bin_weight=DMRInt_weight(bin_method, input_DMR, pd, beta, group1, group2)
}
